# Avoid batching Jaeger spans by Process.