… protocol/signalfx/trace_zipkin_test.go