```bash
$ benchcmp old.txt new.txt

benchmark                              old ns/op     new ns/op     delta
BenchmarkParseJaegerFromRequest-16     39724         32564         -18.02%
```

Remove ParseMapOfJaegerBatchesFromRequest (used only by trace-ingest which will use the new API) and ParseSAPMFromRequest (not used outside this repository).